worldshop/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.2.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
colorama==0.4.6
cryptography==39.0.1
defusedxml==0.7.1
Django==4.1.3
django-allauth==0.52.0
django-appconf==1.0.5
django-cleanup==6.0.0
django-filter==22.1
django-jazzmin==2.6.0
django-modeltranslation==0.18.7
frozenlist==1.3.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httplib2==0.21.0
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
multidict==6.0.4
oauthlib==3.2.2
Pillow==9.3.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.5
pygramtic==0.2.0
PyJWT==2.6.0
pyparsing==3.0.9
pypng==0.20220715.0
python-telegram-bot==20.1
python3-openid==3.2.0
pytz==2022.7.1
qrcode==7.4.2
requests==2.28.2
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.3
typing_extensions==4.4.0
tzdata==2022.6
urllib3==1.26.14
yarl==1.8.2
